Refactor EditTransaction and renamed it to EditTransactionForm
