+++ src/Pages/Edit.js
@@ -1,12 +1,12 @@
-import React from 'react';
-import EditTransactionForm from '../Components/EditTransaction.js';
+import React from "react";
+import EditTransactionForm from "../Components/EditTransactionForm.js";
 
 function Edit() {
-    return (
-        <div>
-            <EditTransactionForm />
-        </div>
-    );
-};
+	return (
+		<div>
+			<EditTransactionForm />
+		</div>
+	);
+}
 
 export default Edit;
